disc-location/pyproject.toml
[project]
name = "disc-location"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "onnxruntime-gpu>=1.23.2",
    "opencv-python>=4.12.0.88",
    "scikit-image>=0.25.2",
    "tensorflow>=2.19.1",
    "tf2onnx>=1.16.1",
]
